requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
celery==5.3.4
redis==5.0.1
docker==6.1.3
psycopg2-binary==2.9.7
gunicorn==21.2.0
dj-database-url==2.0.0
django-cors-headers==4.3.1